Added support for a different port. This at the request of Drew Taylor. The new configuration directive is LISTEN_PORT, and I've added it to both bricolage.conf and bricolage.tmp.conf.in (note that I also set SSL_ENABLE to off by default in the latter file). Now if you set LISTEN_PORT to a value other than 80 -- say 8080 -- then you can access Bricolage via that port, e.g., http://myhost.com:8080/.
I also tried to add an SSL_LISTEN_PORT directive, but this simply didn't work.
The problem is that the browsers save the cookies and send them to the servers
by using the hostname + the port. So if you enable ssl on port 444, then when
you login, the browser won't send the cookie to your reguarl Bricolage server on
a different port, because the domain name won't include the same port number (if
any). The advantage of using 80 and 443 is that the port isn't included in the
domain name, so the cookies get submitted regardless of the port.

So the upshot is that you cannot set LISTEN_PORT to a value other than 80 if you
also have SSL_ENABLE turned on. It throws and error and Bricolage won't start if you
do it. Right now, you can have either this:

LISTEN_PORT = 80
SSL_ENABLE  = on

or this

LISTEN_PORT 8080
SSL_ENABLE  off

or, of course, this:

LISTEN_PORT 80
SSL_ENABLE  off

One other option, however, would be if someone wanted to listen on a port other
than 443 for SSL connections, to not allow them to connect to the regular http
port at all -- everything would be over SSL. This will be slower, but would
allow the user of a different port for SSL. Comments?