Template locating bug fix. For output channels with post_paths, the template often couldn't be found. I added code to make sure to always look under the post_path, too, for every directory. So if you're looking for "/foo/bar/email/story.mc", and "email" is the post path, find_template() will look for the template in this order:
  /foo/bar/email/story.mc
  /foo/email/story.mc
  /email/story.mc

If "email" isn't a post path, find_template() will look for the temlate in this
order:

  /foo/bar/email/story.mc
  /foo/bar/story.mc
  /foo/story.mc
  /story.mc

In reality, template shouldn't be stored under the post_path names, but it was
easier to get the burner to always look there, too, than to fix all the old
template records in the database.

Sam, I'm not sure if this affects how you find HTML::Template templates.