Matt's omnibus patch.  Changes:
-Added a 'Check In/Publish' capabilty for stories and media, 'Check
In/Deploy' for templates.  This is funtionality is one step- just click
the button and it is done (as well as related stories/media that needs
republish).

-Stripped the publish/preview code out of it was repeated in several
places and put it into Bric::Util::Burner as $burner->publish() and
$burner->preview().  Next step to add a function that returns all
related stories and media? (this code is replicated in a few places)

-Added a published_version field to story and media tables.  This is now
used to determine if assets should be republished with
Bric::Asset->needs_publish(), comparing current_version with
published_version.  Old method was simply to see if assets had been
published previously.

-Added (P) graphics to listmanager- no (P) shows if an asset has never
been published, a green (P) if the current version has been published, a
red (P) if the current_version does not match published_version
(needs_publish).

-Story and media profile now shows current and published versions.