Fixed a bug in Bric::Util::Coll that would call the href() method on a class and pass in undefined values as parameters. This will happen if a collection is created for a new object that does not yet have an ID. The fix thus prevents the "= NULL" SQL syntax, which always matches nothing.
The fix itself is to allow an undefined $params argument be be passed to
Bric::Util::Coll and its subclasses. If it's not defined, Bric::Util::Coll
will not execute href(). The way I've used it in Bric::Util::Job is like this:

  Bric::Util::Coll::Resource->new( defined $id ? { job_id => $id } : undef);

This should speed up quite a few things.