Final Output Channel Selection commit. This commit also adds the new "mode" property to Bric::Util::Burner. Here's a more complete list of changes pasted from Bric::Changes:
=item *

Improved the interface for the assocation of output channels in Element
Profile. It now uses a list of output channels rather than a double list, and
a radio button to select the primary output channel. Also, each associated
output channel can be marked as "Enabled" or not, indicating whether a new
asset based on the element will by default be associated with that output
channel for publishing. [David]

=item *

On a related note, Stories and Media can now select which output channels to
be published to on a per-story basis, and the selection sticks with each
version of a an asset. The list of available output channels comes from the
output channels associated with the element on which the story or media asset
is based. [David]

=item *

Bric::Util::Burner objects have a new property, C<mode>. This property
contains an integer value indicating whether a burner object is currently
publishing, previewing, or checking syntax. The value maps to the
conveniently-named constants "PUBLISH_MODE", "PREVIEW_MODE", and
"SYNTAX_MODE". [David]

=item *

The documentation in the Bric::Util::Burner class has been updated and
improved. [David]