Major fixes for the Output Channel Association feature. As you may recall, I added this feature in the last several weeks. However, I discovered a flaw yesterday, and it led me to do a lot of new work on the feature. The result is a better implementation, with some cleaner code, and more reliable generation of asset URIs based on output channel settings.
The problem I ran into was that the URIs of assets weren't always getting
properly formatted for a story. This is because while the element on which
a story is based has a primary output channel, the assets did not. So the
URI could be formatted based on an output channel with which a story wasn't
even associated!

The upshot was that I had to create the concept of a primary output channel
for business assets similar to the primary output channel for elements. The
element on which a business asset is based sets the output channels and
primary output channel on new business assets, and make available to the
business asset a list of output channels to add later, but the URI is now
always built for the primary output channel for a business asset, and the UI
ensures that business assets always have primary output channels.

I had to change the UI, too, but it's actually less complex, now. The output
channel association is handled similar to category association, in its own
section just after category association. Furthermore, I added events for
adding and removing output channels, so you can tell via the log who added and
removed output channels (after a story is created -- the initial list of
output channels is not logged).

Enjoy!