Bric::Util::Coll optimization. I updated del_objs() so that it doesn't automatically load the objects from the database. If the objects are already loaded, it'll use them, but otherwise it won't worry about them. Of course, I also had to update get_objs() so that it checks the del_objs hash and removes them when the collection is populated from the database.
These changes required that I update all of the collection subclasses so that
they use the del_objs property as a hash instead of an array. Also, I removed
the ability for a call to del_objs with no arguments to delete all
objects. This meant that the two spots that exploited this approach had to be
changed to call get_objs() to pass in the list of all objects to delete.

And finally I also added a fairly comprehensive test suite for
Bric::Util::Coll.