OutputChannel Refactoring. Same old story as in the last few days. I rewrote the list() code to add support for a grp_ids property, for finding output channels based on a grp_id, and changed "WHERE IN (SELECT)" clauses to INNER joins.
I've also updated the test suite so that it can run properly when the test
values in the *.tst files aren't loaded into the database. This is in part in
preparation for deprecating those files. Shouldn't be a big deal, since I
think I'm the only person who uses them, anyway. The upshot is that I had to
reorganize things a bit, and I added a user_id() method to Bric::Test::Base,
so that any class can get at and use a legitimate user ID. Might need to
do this for an output_channel ID, too.