Added object caching. Yeah, that's a mouthful, isn't it? I was looking at all the queries executing during a simple publish, and noticed that the Bric::Biz::AssetType::Parts::Data->lookup() was getting called over and over again to retrieve the same objects. For example, this was the query that executed the most times and thus used the most processor time was this one (from the output of bric_dbprof):
### Query   1 #################################################################

 SELECT element__id, name, description, place, required, quantifier,
	autopopulated, map_type__id, publishable, max_length, sql_type,
	active
 FROM at_data
 WHERE id=?
 AND active=?

###############################################################################

 Execution Count: 31
 Total Time:       0.337057 seconds
 Average Time:     0.010873 seconds
 Longest Time:     0.075484 seconds


That's right, it ran 31 times for a total of a third of a second -- for a
lookup! The problem is that the get_data() method of data elements has to
check the SQL type of the data to see if it's a date, and if it is, to return
the properly-formatted date. But to get the SQL type, it has to look up the
Bric::Biz::AssetType::Parts::Data object that describes it. Hence the lookup()
method of that class is executed many times, even though the same type of
field is often looked up (e.g., paragraph).

Unfortunately, there's no easy way around this issue in the current design, so
I decided to apply the 80/20 rule -- get 80% of the benefit with 20% of the
work. This commit is the culmination of my application of the rule.

My thought was to add object caching so that the same object didn't have to be
retrieved from the database every time. At first, I thought to add a
generalized caching for all objects using shared memory between Apache
processes. IPC::MM was a leading candidate because of its speed. However, it
then occurred to me that there might be problems with objects constructed from
other objects, and the chance for an object in the cache to become stale. So,
the 80/20 rule.

Since I noticed that this was a problem for publishing in particular, it
seemed to make sense to create a simple caching architecture to cache objects
for the duration of a request. The Apache pnotes() method makes this very
simple, of course. And it allowed me to still create an API that can be
adapted for use with a more generalized caching architecture in the future. So
that's what I've done here.

So what I did was to add two new methods to Bric.pm: cache_me() and
cache_lookup(). cache_lookup() is a class method, meant to be used by the
lookup() methods to return a cached object. cache_me() is an instance method,
meant to add an object to the cache right after it's retrieved from the
database. I then added calls to these methods in the appropriate places in all
classes that have lookup() methods. Although they now only cache objects for
the duration of a request, they can easily be updated to use a different
caching approach without changing other classes again.

The result now using the Apache pnotes() caching is that the same publish that
executed the at_data table lookup 31 times now does executes it only twice:

### Query  17 #################################################################

 SELECT element__id, name, description, place, required, quantifier,
	autopopulated, map_type__id, publishable, max_length, sql_type,
	active
 FROM at_data
 WHERE id=?
 AND active=?

###############################################################################

 Execution Count: 2
 Total Time:       0.022351 seconds
 Average Time:     0.011175 seconds
 Longest Time:     0.018910 seconds

This cuts down the total time that this query takes up to just 1/50 of a
second for a single story! The same benefit will of course be realized in
similar situations.

Of course I realize that this doesn't necessarily solve the underlying problem
with the performance of some of the Bricolage code. But I hope that it's a
decent stop-gap for now, and along with all the refactoring of the list() code
in many of the classes, to eliminate most of the performance bottlenecks. I'll
do more profiling in the coming weeks to try to find and eliminate other
performance bottlenecks.

All other changes in this commit are mainly POD fixes and documentation fixes.