Fix for Bug # 335. This issue naturally turned out to be larger. I had to update all the classes I could find that look in the member table on or more times in their SELECT statements to always make sure that member.active = 1. I also wrote tests to make sure this stays fixed.
In the process I discovered a few other issues. First, Bric::Biz::Org and
Bric::Biz::Org::Source had a typo in their save() methods that caused them to
deactivate their group memberships every time they were updated. This
obviously isn't a good idea, so I fixed the typo and added
inst/upgrade/1.5.1/activate_all_grp_members.pl to correct the problem.

Second, in those classes that deactivate their group memberships when their
objects are deactivated, they were deactivating *all* group memberships. It
seems like a good idea to me that they should leave their memberships in
permanent groups in place, so I changed the code to do so and added code to
inst/upgrade/1.5.1/activate_all_grp_members.pl to correct existing records
there, too.

But all this revealed that there are some inconsistencies in the way that
group memberships are maintained (or not) by deactivated objects. So I opened
a new bug for this issue so that it's on the radar and can eventually be
addressed.

One last change I've made in this commit was to Bric::SOAP::Util, where there
were problems with elements without subelements. I've corrected this issue and
so all should be well there.

That's enough for now!