Fixed element deletion bug. I'm pretty sure that this bug was only introduced with my recent revisions to the collection API in which I allowed collection members to be deleted without looking up all of the existing members in the database. I wrote tests to re-create the problem, and then added new methods to Bric::Util::Coll::Member to handle the situation. It works when QA_MODE is enabled and when it is not.
This became necessary because of the way that the group API's href() method
sometimes uses member IDs for keys and sometimes the underlying object IDs for
keys. This is to handle groups that can manage different kinds of objects. See
some of the notes I've added to Bric::Util::Coll::Member explaining the
situation.