Added Bric::Biz::Site. It is fully implemented, and the apprpriate upgrade script has also been included. This commit also includes a complete set of tests for this new clas -- over 300 in all. I expect that Bric::Biz::Site::Test and Bric::Biz::Site::DevTest will become the model for a comprehensive set of tests for a single class. Also new in this commit is the style of POD in Bric::Biz::Site. This, too, will become the standard for all Bricolage classes going forward.
The only thing that's not done yet (and for which there are TODO tests) is the
appropriate SQL syntax in list() to exclude sites for user IDs for any user
who is in any group with DENY permission. I hope to get the SQL figured out in
the next few days.

More notes will be added to MultiSite::Implementation shortly.