This adds to GUI to admin part of workflows. It lets you choose site when you create the workflow and then displays it.
There is a bug that makes it impossible to edit a newly created workflow, if you go into the SQL table and set it to active and then try and edit it works. It seems like workflow_id is getting lost somewhere in the callbacks and my mason-fu isn't strong enough to debug it.