Fixes for Bug # 347. There were two issues at work here. The first, related to Media->list() getting called four times, was because Grp.pm was getting a list of classes from its subclasses, and since there are three subclasses of Media, and all three fall back on Media's list() method, it was getting called four times. The solution is a new method on Grp, get_list_classes(), that returns the appropriate list of classes without subclasses.
The second issue is that the new large fancy queries on assests weren't
sorting on the "id" column as the last sort column. This caused the columns to
be returned such that rows for the same object weren't all concurrent. I've
fixed this by changing order_by() in Bric::Util::DBI to *always* ultimately
sort by "id". I've also removed "id" as a supported argument to the "Order"
parameter, since "id" is *always* an order by column now. I updated the
relevant tests and they all pass.