Gutted Bric::Biz::Keyword. I slipped a sharp knife up into its body and drained it of its contents. Then I filled it back up with the latest Bricolage goodness.
This is a complete redesign and refactoring of the keyword code. Non-standard
methods have been deprecated. Associations with categories, media, and stories
are now handed by a collection object instead of directly. Non-standard
methods in Category.pm have been deprecated and replaced with standard
methods. The keyword association methods in Business.pm can now takea list of
keywords instead of just an array reference of keywords as their arguments.

I also added a new exception, Bric::Util::Fault::Error::Undef, which is thrown
by Keyword.pm and Site.pm when required values have no value. Also, keywords
are now groupable, and all are members of the "All Keywords" group, which
will make it easier to create a keyword manager in the UI.

I added a full test suite for keyword objects (with more tests to come after
the next merge from rev_1_6), and updated some of the site tests.