Fixes for Bugs 340 and 370. I've added a new bricolage.conf directive, STORY_URI_WITH_FILENAME, which allows stories to have the filename in URIs. This makes it a bit easier to have stories in the same directory but with different file names (as when you use slugs for the file name).
I've also moved the creation of the httpd.conf and bricolage.conf (and
install.db) files up so that they're created during `make` rather than `make
install`. They get put into a special directory, bconf, and Bric::Config knows
to look there if tests are being run (that is, when $ENV{BRIC_TMP_DIR} is
set). I've also changed Bric::App::Cache and Bric::App::Session so that they
don't try to `chown` any files are directories if tests are running.

Tests of the updates to the installer code would be most welcome.