Ancient bug fixed. This must be the week of the really old and obscure bugs! This is the third one this week!
Fixed an obscure bug where an element with a field element and a container
element with the same ID could cause an error when one tries to edit the
container element and Bricolage tries to load the field element,
instead. Embarrassingly discovered in the middle of a presentation to the
London Perl M[ou]ngers.