Various install & upgrade fixes.
The "multi" installation method now defaults the "Bricolage Root Directory"
setting to /usr/local/bricolage instead of "NONE". This seems to be less
annoying to people.

During make upgrade, the upgrade scripts are now run _before_ the new APIs are
installed, so that there are no conflicts when using the existing API to make
changes.

Creation and installation of man pages can now be avoided during make upgrade
as well as during make.

Man pages will now be installed in the correct subdirectory of $BRICOLAGE_ROOT
when using the "multi" install method.