Trying to implement David's suggestion of "redirecting" on errors in Handler.pm the same way as in PreviewHandler.pm; that is,
   sub handle_err {
       my ($r, $err);
       ...
       $r->uri(ERROR_URI);
       $r->filename(ERROR_FILE);
       $ah->handle_request($r);
   }

instead of ending with

    return $interp->exec(ERROR_URI, fault => $err,
                         __CB_DONE => 1, more_err => $more_err);

so that we get rid of an undefined $m.
   But I'm not doing this right somehow. I'm not sure if
it's just because I'm returning the wrong Apache constant
or what. To demonstate the problem, login and try to create
a new Media (MEDIA->New Media, Create); when you click Create,
the page will not be redirected to 500.mc but you will find
that your error_log has an error/stack trace (unrelated error
which I don't know the cause of yet). I'm returning 'OK' so maybe
that's not right (I thought it was though, to say that we
handled this part.) It might just be that using $gah (hehe)
doesn't make sense for some reason.