Site error and localization changes. This commit does two things. First, it removes the eval in the site callback, since that is now handled by the callback exception handler. Yay! Gotta love simplification.
The second thing it does is change the behavior of some calls to add_msg() so
that they're no longer passing in HTML entities. Instead, we use literal
quotation marks. This is so that the localizations can change them. For
example, I changed the de_de.pm localization to use << and >> instead of ", as
those are the quotation marks used for German. (Ignore the fact that de_de.pm
is still actually Portuguese).

And finally, to make this work, header.mc now escapes the HTML stored via
add_msg(). This is good, as it allows other future interfaces to use the
add_msg() messages in their own ways.

What this means is that the localization audit is now required. I'll post a
message to the list shortly outlining the bare mininum changes we need to make
to all calls to add_msg().