Fix for Apache 1.3.28. It turned out to be our problem! At least so it appears. The solution was to get rid of our own database connection caching and use DBI's connect_cached(), instead. This is also likely more efficient, too. I also fixed an issue where Bric::Util::DBI could create inconsistent transaction states. These might have been part of the problem though I'm not sure how.
And finally, with all these changes to Bric::Util::DBI, I decided to clean it
up a bit. It's now much more readable, and uses the new exception throwing
syntax. I also removed documentation of the $DEBUG argument to some of the
functions, and then changed the functions so that they pass @_ directly to the
DBI methods. This caused some problems for those classes that were still
passing in $DEBUG arguments, so I stripped all of those out. I also fixed
those few classes that were omitting a second argument to prepare() and
friends.

All this leads to a rather big commit, but it's not such a huge change,
really, and definitely for the better!