Added preview widget. Also combined the "Find" components into a single dhandler. From Bric::Changes:
=item *

All preview links are now generated by a single widget. This widget adds the
story or media URI to the "title" attribute of the link tag (which is modern
browsers will automatically work as a roll-over tooltip), makes the story or
media URI copyable (by relying on JavaScript to actually open a new window to
the preview), and manages selecting an output channel in which to preview a
story. [David]

=item *

Added ability to preview stories in a selection of output channels in the View
(read-only) story profile. [David]

=item *

Combined the asset "Find" managers into a single component, thus eliminating a
lot of redundant code. [David]