Added DTD validation action. This is a new distribution action that you can use to validate your output files against a DTD (or whatever XML::LibXML understands). I've also simplified the way actions are subclassed, so you don't actually have to edit Action.pm to add a subclass now, but you do have to load your actions in Bric::App::Handler (and in the test suite).
I've also added code to Bric::Test::Base to tie off STDOUT, STDERR, and
STDIN. This is so you can take more control of output in your tests. See the
new read_stdout(), read_stderr(), and print_stdin() methods for more
information. I used the ability to read STDERR in the DTDValidate DevTest.
This is because DTD validation can be so slow (XML::LibXML fetches the DTD
over the Internet, if possible) that I was having it output a status message
for each file it validates.

This patch is not as big as it looks -- most of it is the XHTML Strict DTD,
which I've included in its entirety in order to be able to write tests that
actually work. If anyone who knows more about XML that I do would like to
replace my examples with a much simpler DTD, then feel free!

Enjoy!