Changes due to single transaction upgrade scripts. First, permissions weren't geting properly set; the permissions weren't granted to the root user at all. This is because, I think, we were granting permission to the bricolage Pg user in the same transaction, so PostgreSQL didn't bother to give them to the user that created the tables.
The second issue is that it appears that one cannot add a table and then alter
it (add constraints) within the same transaction. So from now on, constraints
will either have to be added as part of adding the table, or else in a
separate upgrade script, as I've done here.