Fix for Bug # 493. I've added a constraint so that it's impossible to associate an element with a site without also having an output channel association. And since the UI enforces associating at least one site with each top-level element, it's very difficult to create a top-level element without a at least one site and output channel association.
I also added code so that the select list of sites includes only sites that
are associated with output channels. I needed to add a bit of ugly interface
to Bric::Biz::Site's list() method to make this work efficiently (that is, a
valid value for the new "output_channel_id parameter is "not null"), but work
it does.

So now when you go to create a story or a media document, all of the story
type or media type elements listed will have output channels associated with
them.