Major upgrade architecture overhaul. Well, not that major. Mainly all I've done is to eliminate the test_sql() function in the upgrade library, replacing it with the more intelligent functions test_table(), test_column(), test_constraint(), and test_index(). Upgrade scripts are no longer allowed to trigger SQL errors without failing. This is because the old test_sql(), when it triggered a failure, it ruined the transaction and thus prevented any succeeding do_sql() calls to fail. The new functions check the PostgreSQL schema tables for the needed information, instead.
I've also gone back and installed Bricolage 1.4.0, and made some tweaks to
some of the upgrade scripts to make sure that they can all be run
successfully. So anyone with 1.4.0 or later should be able to successfully
upgrade without too much trouble. The changes mainly involved making sure that
any script that creates a table doesn't add indexes or contraints to that
table. I've added new scripts to do this. This is because, it appears, you
can't add a table and then add indexes or contraints to that table in the same
transaction. Go figure.