Two more big query changes. First, I've added an aggregate that takes a column of group IDs and, for any single story, media, or template object, keeps them in a single row. Thus, the data for any one story, media, or template object is now returned in single rows instead of potentially many rows. This adds a little more work to the database back end, but greatly cuts down on the overhead of fetching data from the database in application space. Over time, I will probably add this optimization to other Bricolage clases, as well.
Second, now that all stories, media, and temlates are returned in single rows,
I was able to move the Offset and Limit parameters to the database back-end,
too. Thus if you're doing a query with Offset => 1000, the API no longer has
to fetch 1000 objects' worth of data before constructing the remaining
objects. It can just construct the objects the database returns, instead. This
makes using Offest and Limit far more efficient.