Fix for upgrade permissions problem. We just have to be sure to always load bric_upgrade.pm _before_ loading Bric::Util::DBI in upgrade scripts. Otherwise, the upgrade scripts run as the Bricolage database user rather than the PostgreSQL super user.
Next, I discovered the need to load Cache::FileCache at startup time. This is
because the error genereated by Storable.pm about being unable to load
Log/Agent.pm is only triggered when it was loaded at runtime (via require or
eval).

Finally all these changes unveiled a number of other problems with the URI
uniqueness upgrade scripts, related to problems using either the previously
installed versions of Story.pm or Media.pm or the new ones in the installation
source directory, neither of which actually work. So I had to rewrite the
script to do it all without those classes. In the process, I realized that
they hadn't been paying attention to the STORY_URI_WITH_FILENAME
bricolage.conf diretive. So I added code to try to find the existing
bricolage.conf file (the upgrade scripts always use the one in the sources)
and parse out that diretive's settings.

I moved the upgrade scripts into the 1.7.3 upgrade directory so as to fix
those installations that had run by the old upgrade scripts and probably
failed. So now, the upgrade process to 1.7.3 should be quite smooth. It's
working well for me now, at any rate!