fix for bug #573
When executing a template component, the request class should not be
H::M::Request::ApacheHandler, as we don't really need that class's
special handling of output (which we trap via our own out_method), nor
do we want to handle "top level comp not found" errors by returning a
404.

Now, we trap HTML::Mason::Exception::TopLevelNotFound exception
ourself, and treat them as a generic error.

If we rethrow them, it percolates to the Mason request executing the
bricolage component (/workflow/profile/preview/dhandler) which is
bizarre and confusing.