Completely documented element classes. It's probably not perfect, but it's a far cry from the incomplete and inaccurate documentation that was there.
A few other things, in this commit, too:

Refactored quite a bit of the code in the element classes. Renamed the methods
with "tile" in their names to use "element" instead (but kept the old ones
around as aliases, since they're used throughout the UI). Added a few methods
to make the interface more complete.

Modified the "get_elements()" method of
Bric::Biz::Asset::Business::Parts::Tile::Container to take an optional list of
key name arguments, and to return only the subelements with those key names.

Added the "get_data_elements(): method to
Bric::Biz::Asset::Business::Parts::Tile::Container. This method functions
exactly like "get_containers()" except that it returns data element objects
that are subelements of the container element. It also takes an optional list
of key name arguments, and, if passed, will return only the subelements with
those key names.

The template profile's "cheat sheet" of the subelements of an element now
correctly display subelement key names instead of munging element names, as
was required before version 1.7.0. I also fixed the display so that commas
appear where they should in those lists.