Added language-specific style sheets. I realized that some languages would require this. The example here now is for Uyghur (ug.css), for which the display of content must be right-to-left instead of left-to-right. But other languages may also need to make changes (such as increase font sizes or use different fonts). This also allows individual Bricolage users to tweak CSS for their own language needs.
I also added optional IDs and classes to table top divs, so that it's easier
to select content fields in CSS. I added the "fieldcontent" class to title and
description and container element content displays for when they're not output
in form fields. See ug.css for examples of these in use.

This commit also localizes a couple of strings that weren't localized before.