Fixes for Bug # 707.
When Bricolage discovers a document or template on a desk but not in a
workflow, it now places it in a workflows associated with the same site as the
document or template is associated with. Thanks to Serge Sozonoff for the
spot!

Aliases of Image, Audio, or Video media documents no longer remain stuck on
desks. Reported by Serge Sozonoff.

The C<has_member()> method of Bric::Util::Grp is now faster when searching for
a specific group member.