Fix for Bug # 675. It actually involved a bit more work that I expected, but it's all good. First, I allowed formerly secret group classes to be non-secret. This means that preference, alert type, and group groups can now be created and managed in the UI. Then I added the group doubleList manager interface to those profiles that were missing it: Preferences, Groups, Alert Types, Element Types, Keywords, and Contributors.
Keyword groups actually weren't listing their members, but the do now.

Contributor groups are the same class as Contributor Types. This is nasty, but
can't be avoided. I added workaround code to change the display name of the
class in the UI. One hack was to provide an optional third array element to
the "addition" parameter to listManager.mc that contains the name of the
object to be added. Yeech. Another hack was to make sure that new contributor
types continue to be created as secret groups, so that they can be
distinguished from plain old contributor groups. Hacks all around!

And finally, I made the table of contacts in the alert type profile nicer, with
the same border colors as similar tables in the UI.

Oh, and the display and plural names of the group group class needed to be
updated, and I added a script to make sure that existing contributor groups
are all secret groups.