Added RELATED_MEDIA_UPLOAD bricolage.conf directive that allows editors to upload media files directly into an element profile if they have permission to create media and to access a media workflow. This makes it a much simpler process for editors to quickly associate media with story documents. The uploaded media files will be created based on the media type element selected from a select list, and given the same category, cover date, and priority as the current document.
Also added convenient find_workflow() and find_desk() functions to
Bric::App::Util. These facilitate finding a given workflow in a given site
that the user has a given access to, and a desk in a given workflow and for
which they have a given permission to access the assets on that desk.