Convert all "tabs" to XHTML and CSS.
These are all of the section headers with curved corners.  It now uses two images with a white 
background and transparent foreground to create the curved effect, replacing 74 images with 2! The 
sections also now resize to fit their container in anticipation of a fully fluid-width layout, but are 
set for now at 580 pixels until the rest is done.