Fix for Bug # 795. When Bricolage notices that a document or template is not in workflow or on a desk when it should be, it is now more intelligent in trying to select the correct workflow and/or desk to put it on, based on current workflow context and user permissions.
I've also added output the the Apache log consisting of the 20 most
recent events on an asset that has lost track of what workflow and/or
desk it is on. This is to provide debugging information that might
help us to track down when an asset forgets its workflow and/or desk,
so that we might eventually be able to eliminate the code that picks
up on this and corrects it.