Fix for Bug # 803 and then some.
First, I added the PARAM_ANYWHERE_MAP constant to Story, Media, and
Formatting. Each of the WHERE parameters in this hash has two parts used for
when the ANY() operator is used to pass an argument for the parameter to
list() or list_ids(). The first part is the part of the WHERE clause that must
be included for all of the items to be matched agains. The second part is the
single WHERE expression to be used for each instance of each item in the ANY
argument. This keeps Bricolage from creating redundant WHERE expressions that
can cripple the database.

In the process of working on this, I noticed and corrected a few other
things. I noticed that the documentation of the parameters to
Formatting->list() were woefully inadequate, so I made them complete.

Unrelated to the list() issue, I noticed that the constraints on the "job"
table and friends were missing, and likely have been missing since 1.8.0, when
Mark moved Job from Dist to Util. So I restored them and then added an upgrade
script to restore them.