Fix for Bug # 827.
Added the published_version parameter to the list() methods of the story,
media, and template classes. This parameter forces the search to return the
versions of the assets as they were last published, rather than the most
recent version. This will be most useful to those looking up other documents
in templates and publishing them, as a way of avoiding pulling documents out
from other anyone who might have them checked out!