Clone and list() improvements.
Added support for ANY() function to pass multiple possible values for a
given parameter to the list() method in Bric::Util::Grp. More classes will
support it soon!

Added support for associating secondary categories and output channels when
cloning a story.

Added support for adding secondary output channels to a new clone by selecting
a group of output channels. All output channels in that group that are allowed
by the story element will be associated with the clone. This is an easy and
accurate way to add a bunch of output channels to a clone at once. Users must
have READ access to each listed output channel group.

Also cleaned up latest Bric::Changes listings.