Fixed nasty bug in Bricolage 1.8.4. Much to my embarrassment, just after releasing 1.8.4, I was updating the Kineticode site in the Kineticode Bricolage, which I had just upgraded to Bricolage 1.8.5. Much to my chagrin, I discovered that it was no longer possible to preview un-checked-out stories without getting an error. D'oh! Fortunately, I hadn't announced the release anywhere (I'd published the announcement on the Bricolage site, but pulled it after a few hours). And now I've fixed the bug.
The change in the behavior of the "checked_out" parameter has had deeper
ramifications than I'd realized. The upshot is that I had to add a new
parameter, "checked_in", to emulate the behavior of when the old version of
"checked_in" was false. The "checked_in" parameter, when passed a true value,
will cause the story, media, and template classes to return the instances of
stories as they were last checked-in. This means that checked-out stories are
returned, but any changes made since the checkout are not returned. This is
the desired behavior for "Viewing" stories, media, or templates, or previewing
stories and media. I've updated all of the preview and view links to use this
parameter instead of the checked_out parameter and removed an ugly hack in
Bric::Biz::Asset that had made this all work in 1.8.3, and now all is well
again.

I also updated a spot in Bric::Util::DBI where the "user__id" parameter was
forcing a search for checked-out assets. I made it so that the equivalent
"user_id" parameter does the same thing. That led to test failures, and while
I was fixing them, I added new tests for this behavior, as well as for the new
behaviors of the "checked_out" and "checked_in" parameters.

I've also bumped up the version number, and I will release 1.8.5. shortly.